Use Int for IRK cache counters on 32-bit Julia (#147)
* Use Int for IRK cache counters on 32-bit Julia.

length(u0) and related counters are Int32 on i686; Int64-typed cache
fields caused MethodError when constructing IRKGL_SIMD_Cache. Also
accept Integer indices in VecArray linear indexing.


* Use Ref{Int} for IRK step_number on 32-bit Julia.

Array{Int,0} still constructed as Array{Int64,0} in the x86 CI
MethodError path; Ref{Int}(0) keeps the counter native-width.


* Update alloc_tests to construct step_number as Ref{Int}.

Matches the IRKGL cache field type after the 32-bit counter fix.


* Avoid non-public Base.RefValue in IRK cache field types.

ExplicitImports QA rejects qualified access to Base.RefValue; use
typeof(Ref(0)) so step_number stays a native-Int Ref on every arch.


---------

+++ src/simd/VecArray_def.jl
@@ -32,14 +32,14 @@ end
 
 ## getindex and setindex! implementations for linear indexing
 
-@inline function Base.getindex(v::VecArray{s, T, dim}, k::Int64) where {s, T, dim}
+@inline function Base.getindex(v::VecArray{s, T, dim}, k::Integer) where {s, T, dim}
     j = s * (k - 1)
     #    Vec{s, T}(NTuple{s, T}(@inbounds v.data[is + j] for is in 1:s))
     return Vec{s, T}(ntuple(is -> @inbounds(v.data[is + j]), s))
 end
 
 @inline function Base.setindex!(
-        v::VecArray{s, T, dim}, vk::Vec{s, T}, k::Int64
+        v::VecArray{s, T, dim}, vk::Vec{s, T}, k::Integer
     ) where {s, T, dim}
     j = s * (k - 1)
     @inbounds for is in 1:s
@@ -48,7 +48,7 @@ end
     return nothing
 end
 
-@inline function Base.setindex!(v::VecArray{s, T, dim}, vk::T2, k::Int64) where {s, T, T2, dim}
+@inline function Base.setindex!(v::VecArray{s, T, dim}, vk::T2, k::Integer) where {s, T, T2, dim}
     vk_ = convert(T, vk)
     j = s * (k - 1)
     @inbounds for is in 1:s
